Add files via upload
This code uses a trading strategy based around Divergences on the MACD, RSI, and EMA movements to enter.
Exit strategy is based around setting the stoploss as the previous day's H/L and exitting the moment a bearish candlestick formation is identified. The Risk to Reward ratio is currently set at 2, but can be changed according to convenience.